Rename clicks.ts to Clicks.ts
