+++ src/provider/view/components/Controls/Button/Button.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {usePreventMouseDownPropagating} from '../../../hooks/clicks';
+import {usePreventMouseDownPropagating} from '../../../hooks/Clicks';
 import './Button.scss';
 
 interface Props {
